+++ ms-user-manager/src/main/java/com/postech30/msusermanager/entity/User.java
@@ -16,6 +16,9 @@ public class User {
     private String password;
 
     @ManyToMany(fetch = FetchType.EAGER,cascade = CascadeType.PERSIST)
+    @JoinTable(name = "users_roles",
+            joinColumns = @JoinColumn(name = "user_id"),
+            inverseJoinColumns = @JoinColumn(name = "role_id"))
     private List<Role> roles;
 
 }
